refactor exercise handling
