--- backend/src/main/java/org/conexaotreinamento/conexaotreinamentobackend/api/exception/BusinessException.java
@@ -1,15 +1,0 @@
-package org.conexaotreinamento.conexaotreinamentobackend.api.exception;
-
-import lombok.Getter;
-import org.springframework.http.HttpStatus;
-
-@Getter
-public class BusinessException extends RuntimeException{
-    private final HttpStatus status;
-
-    public BusinessException(String message, HttpStatus status) {
-        super(message);
-        this.status = status;
-    }
-
-}
